requirements.txt
google-cloud-bigquery
lxml
requests
